Fix test steps being ended incorrectly due to spurious 'runner:result' events
There are two cases that I've observed where code is expecting a particular test step to be the suite's currentStep and is incorrect.
1. When a test fails, and wdio is configured to save a screenshot, a new 'runner:command' event is sent to fetch the screenshot.  This may happen before the previous 'runner:command' has received its 'runner:result' event (for example, if the test failed due to a timeout).  In this case, we see the following ordering of events:

runner:command#1 - last browser command
runner:command#2 - screenshot
runner:result#1 - last browser command - calls endStep() for screenshot step
runner:command#2 - screenshot - calls endStep() for last browser command

In practice, this is mostly harmless; data gets attached to the wrong step, but life goes on.

2. When a test times out, the 'test:fail' event may be triggered before the last wdio browser command actually completes and sends its 'runner:result' request.  This can result in the following sequence:

runner:command
test:fail
test:start
runner:result

Currently, in practice, this is also mostly harmless; 'test:fail' ends the steps appropriately, and I don't believe I've ever observed a second 'runner:command' coming in before the previous test's 'runner:result'.

However, when implementing webdriverio/wdio-allure-reporter#37, this becomes a significant problem - I *have* observed an 'allure:startstep' event coming in before the previous test's 'runner:result', and this results in the entire test's step reporting tree being totally incorrect.

This patch mitigates this problem by always checking which test step it is attempting to end before ending it. It postpones ending the step if it's found on the stack but not at the top, and it ignores ending the step if it is not in the stack at all.